requirements.txt
alembic
certifi
cffi
charset-normalizer
clearbit
click
colorama
db
dnspython
email-validator
chardet
Flask
Flask-Dance
Flask-DebugToolbar
Flask-Login
Flask-Migrate
Flask-OAuth
Flask-OAuthlib
Flask-SQLAlchemy
Flask-WTF
Flask-Admin
Flask-Bcrypt
Flask-Ext
Flask-JWT
Flask-JWT-Extended
Flask-RESTful
greenlet
iniconfig
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Mako==1.0.7
MarkupSafe==2.0.1
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==21.0
Pillow==8.3.1
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
urllib3==1.26.6
versiontools==1.9.1
Werkzeug==2.0.1
WTForms==2.3.3
requests-oauthlib==0.8.
SQLAlchemy-Utils==0.33.2
URLObject==2.4.3
wincertstore==0.2
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
flask-marshmallow==0.14.0
aniso8601==9.0.1
antiorm==1.2.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30